fix(controller) : delete product wich doesnt have items

+++ pos-be/src/controller/inventories/products/delete.ts
@@ -4,8 +4,20 @@ import { cloudinaryImageDestroy } from "../../../utils/cloudinary.js";
 
 export const deleteProduct = async (req: Request, res: Response) => {
     try {
-        const { id } = req.params
-        const { businessId } = req.body
+        const { businessId, id } = req.params
+        const used = await prisma.item.count({
+            where: {
+                productId: id ?? {}
+            }
+        });
+
+        if (used) {
+            return res.status(400).json({
+                message: "Tidak bisa menghapus. Produk sudah memiliki riwayat penjualan.",
+                code : res.statusCode
+            });
+        }
+
         const product = await prisma.product.delete({
             where: {
                 id: id as string,
@@ -16,11 +28,13 @@ export const deleteProduct = async (req: Request, res: Response) => {
         await cloudinaryImageDestroy(product.id)
 
         return res.json({
-            message: 'product has been updated',
+            message: 'product has been deleted',
             data: product
         })
     } catch (error) {
-       
+        console.log(error);
+
+
         return res.status(500).json({
             message: 'error on server when delete product',
             code: res.statusCode,
